backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
ultralytics==8.1.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
httpx==0.25.2
